+++ vs12/Samples/06_RayTracing/Lightsource.h
@@ -10,13 +10,13 @@
 class Lightsource
 {
 public:
-	Lightsource(const Bow::Core::Vector3<double>& location, const Bow::Core::ColorRGB& color, double intensity); // Erwartet Position, Farbe und Intensitaet
+	Lightsource(const Bow::Core::Vector3<float>& location, const Bow::Core::ColorRGB& color, float intensity); // Erwartet Position, Farbe und Intensitaet
 
-	Bow::Core::Vector3<double>& getLocation(); // Position
+	Bow::Core::Vector3<float>& getLocation(); // Position
 	Bow::Core::ColorRGB& getColor(); // Farbe (Lichtstaerke)
 
 
 protected:
-	Bow::Core::Vector3<double> mLocation; //! Position
+	Bow::Core::Vector3<float> mLocation; //! Position
 	Bow::Core::ColorRGB mColor; //! Farbe (Lichtstaerke)
 };
