+++ vs12/Samples/06_RayTracing/Triangle.h
@@ -12,15 +12,15 @@
 class TriangleObject : public IObject
 {
 public:
-	TriangleObject(const Bow::Core::Triangle<double> &triangle, int matIndex); // Erwartet Eckpunkte und Material
+	TriangleObject(const Bow::Core::Triangle<float> &triangle, int matIndex); // Erwartet Eckpunkte und Material
 
 	virtual bool closestIntersection(Intersection& hit); // Am naechsten gelegener Schnittpunkt
-	virtual bool anyIntersection(Bow::Core::Ray<double>& ray, double maxLambda); // Schnittpunkttest mit einem Strahl
+	virtual bool anyIntersection(Bow::Core::Ray<float>& ray, float maxLambda); // Schnittpunkttest mit einem Strahl
 	virtual void calcNormal(Intersection& hit); // Normale am Schnittpunkt berechnen
 
 protected:
-	Bow::Core::Vector3<double> m_PlaneBase;
-	Bow::Core::Vector3<double> m_PlaneVector1;
-	Bow::Core::Vector3<double> m_PlaneVector2;
-	Bow::Core::Vector3<double> m_Normal;
+	Bow::Core::Vector3<float> m_PlaneBase;
+	Bow::Core::Vector3<float> m_PlaneVector1;
+	Bow::Core::Vector3<float> m_PlaneVector2;
+	Bow::Core::Vector3<float> m_Normal;
 };
